// FrontEnd/src/app/api/model/sc-metadata-dto.ts
/**
 * Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ScMetadataDto { 
    uploader: string | null;
    thumbnail: string | null;
    thumbnails: Array<string> | null;
    title: string | null;
}